admin/requirements.txt
django==4.0.4
django-model-utils==4.3.1
requests~=2.26.0
gunicorn==20.1.0
psycopg2-binary==2.9.5